AIAI-70 refactor: separate internals form re-usable parts
- clarify that our tools are currently focusing on SubProcesses
