+++ axon-ivy-ai-test/src_test/com/axonivy/utils/ai/tools/test/TestIvySubProcessToolSpecs.java
@@ -8,7 +8,7 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 
 import com.axonivy.utils.ai.test.Person;
-import com.axonivy.utils.ai.tools.JsonToolParamBuilder;
+import com.axonivy.utils.ai.tools.internal.JsonToolParamBuilder;
 
 import ch.ivyteam.ivy.application.IProcessModelVersion;
 import ch.ivyteam.ivy.environment.IvyTest;
@@ -21,7 +21,7 @@ import dev.langchain4j.model.chat.request.json.JsonStringSchema;
 
 @IvyTest
 @SuppressWarnings("restriction")
-class TestIvyToolsSpec {
+class TestIvySubProcessToolSpecs {
 
   @Test
   void complexParams() {
