+++ axon-ivy-ai-test/src_test/com/axonivy/utils/ai/tools/test/TestIvySubProcessToolsProvider.java
@@ -11,9 +11,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 import com.axonivy.utils.ai.connector.OpenAiServiceConnector.OpenAiConf;
-import com.axonivy.utils.ai.tools.IvyToolExecutor;
-import com.axonivy.utils.ai.tools.IvyToolSpecs;
-import com.axonivy.utils.ai.tools.IvyToolsProvider;
+import com.axonivy.utils.ai.tools.IvySubProcessToolsProvider;
+import com.axonivy.utils.ai.tools.internal.IvySubProcessToolExecutor;
+import com.axonivy.utils.ai.tools.internal.IvySubProcessToolSpecs;
 
 import ch.ivyteam.ivy.bpm.exec.client.IvyProcessTest;
 import ch.ivyteam.ivy.environment.AppFixture;
@@ -29,7 +29,7 @@ import dev.langchain4j.model.chat.response.ChatResponse;
 import dev.langchain4j.service.AiServices;
 
 @IvyProcessTest(enableWebServer = true)
-class TestIvyToolsProvider {
+class TestIvySubProcessToolsProvider {
 
   @RegisterExtension
   LoggerAccess log = new LoggerAccess(LoggingHttpClient.class.getName());
@@ -45,7 +45,7 @@ class TestIvyToolsProvider {
     var model = aiMock();
     UserMessage init = UserMessage.from("Help me, my computer is beeping, it started after opening AxonIvy Portal.");
 
-    List<ToolSpecification> ivyTools = IvyToolSpecs.find();
+    List<ToolSpecification> ivyTools = IvySubProcessToolSpecs.find();
     ChatRequest request = ChatRequest.builder()
         .messages(init)
         .toolSpecifications(ivyTools)
@@ -57,7 +57,7 @@ class TestIvyToolsProvider {
         .isNotEmpty();
 
     var results = aiMessage.toolExecutionRequests().stream()
-        .map(IvyToolExecutor::execute).toList();
+        .map(IvySubProcessToolExecutor::execute).toList();
 
     ChatRequest request2 = ChatRequest.builder()
         .messages(Stream.concat(Stream.of((ChatMessage) init, aiMessage), results.stream()).toList())
@@ -77,7 +77,7 @@ class TestIvyToolsProvider {
   void chatAgentic() {
     var supporter = AiServices.builder(SupportAgent.class)
         .chatModel(aiMock())
-        .toolProvider(new IvyToolsProvider())
+        .toolProvider(new IvySubProcessToolsProvider())
         .build();
     String chat = supporter.chat("Help me, my computer is beeping, it started after opening AxonIvy Portal.");
     System.out.println(chat);
