+++ axon-ivy-ai/src/com/axonivy/utils/ai/tools/internal/IvyToolsProcesses.java
@@ -1,4 +1,4 @@
-package com.axonivy.utils.ai.tools;
+package com.axonivy.utils.ai.tools.internal;
 
 import java.util.EnumSet;
 import java.util.List;
@@ -10,16 +10,16 @@ import ch.ivyteam.ivy.process.rdm.IProcessManager;
 import ch.ivyteam.ivy.process.rdm.IProjectProcessManager;
 
 @SuppressWarnings("restriction")
-class IvyToolsProcesses {
+public class IvyToolsProcesses {
 
   private final IProjectProcessManager processes;
 
-  IvyToolsProcesses(IProcessModelVersion pmv) {
+  public IvyToolsProcesses(IProcessModelVersion pmv) {
     IProcessManager man = ch.ivyteam.ivy.process.rdm.IProcessManager.instance();
     this.processes = man.getProjectDataModelFor(pmv);
   }
 
-  List<CallSubStart> toolStarts() {
+  public List<CallSubStart> toolStarts() {
     // TODO: craft with APIs that are public accessible
     return processes.searchAll(EnumSet.of(ProcessKind.CALLABLE_SUB))
         .tag("tool")
